requirements.txt
Django==4.1.1
djangorestframework==3.14.0
numpy==1.23.3
tensorflow==2.10.0
